+++ exporters/variables-scss/src/content/generator.ts
@@ -1,120 +1,52 @@
 import { TokenGroup, Token, TokenType } from '@supernovaio/sdk-exporters';
-import { exportConfiguration } from '../index';
-import { generateCssObject } from './generators/cssGenerator';
-import { dimensionTokenToCSS } from './generators/dimension';
+import { createFileWithContent } from './generators/fileGenerator';
 
-type TokenHandler = (
-  token: Token,
-  mappedTokens: Map<string, Token>,
-  tokenGroups: TokenGroup[],
-  withParent: boolean,
-) => string | null;
-
-const addDisclaimer = (content: string): string => {
-  if (exportConfiguration.generateDisclaimer) {
-    return `/* This file was generated by Supernova, don't change manually */\n${content}`;
-  }
-
-  return content;
-};
-
-const tokensToCSS = (
-  tokens: Token[],
-  handler: TokenHandler,
-  mappedTokens: Map<string, Token>,
-  tokenGroups: Array<TokenGroup>,
-  withParent: boolean,
-): string => {
-  return tokens
-    .map((token) => handler(token, mappedTokens, tokenGroups, withParent))
-    .filter(Boolean)
-    .join('\n');
-};
-
-export const createFileWithContent = (
-  tokens: Token[],
-  mappedTokens: Map<string, Token>,
-  tokenGroups: Array<TokenGroup>,
-  fileName: string,
-  tokenTypes: TokenType[],
-  groupNames: string[],
-  withCssObject: boolean,
-  withParent: boolean = false,
-) => {
-  let cssTokens = '';
-  let cssObject = '';
-
-  tokenTypes.forEach((tokenType) => {
-    groupNames.forEach((group) => {
-      const filteredTokens = tokens.filter(
-        (token) => token.tokenType === tokenType && token.origin?.name?.includes(group),
-      );
-
-      cssTokens += tokensToCSS(filteredTokens, dimensionTokenToCSS, mappedTokens, tokenGroups, withParent);
-      cssTokens += '\n\n';
-
-      const tempCssObject = generateCssObject(filteredTokens, mappedTokens, tokenGroups, withParent);
-      if (tempCssObject !== null) {
-        cssObject += tempCssObject;
-      }
-    });
-  });
-
-  let content = withCssObject ? `${cssTokens}${cssObject}` : cssTokens;
-  // Remove extra blank lines
-  content = content.replace(/\n{3,}/g, '\n\n');
-
-  return {
-    fileName,
-    content: addDisclaimer(content),
-  };
-};
+const filesData = [
+  {
+    fileName: '_spaces.scss',
+    tokenTypes: [TokenType.dimension],
+    groupNames: ['Spacing'],
+    withCssObject: true,
+    withParent: false,
+  },
+  {
+    fileName: '_radii.scss',
+    tokenTypes: [TokenType.dimension],
+    groupNames: ['Radius'],
+    withCssObject: true,
+    withParent: false,
+  },
+  {
+    fileName: '_borders.scss',
+    tokenTypes: [TokenType.dimension],
+    groupNames: ['Border'],
+    withCssObject: false,
+    withParent: true,
+  },
+  {
+    fileName: '_other.scss',
+    tokenTypes: [TokenType.dimension, TokenType.string],
+    groupNames: ['Grid', 'Container', 'Breakpoint'],
+    withCssObject: true,
+    withParent: true,
+  },
+];
 
 export const generateFiles = (
   tokens: Array<Token>,
   mappedTokens: Map<string, Token>,
   tokenGroups: Array<TokenGroup>,
 ) => {
-  return [
-    createFileWithContent(
+  return filesData.map(({ fileName, tokenTypes, groupNames, withCssObject, withParent }) => {
+    return createFileWithContent(
       tokens,
       mappedTokens,
       tokenGroups,
-      '_spaces.scss',
-      [TokenType.dimension],
-      ['Spacing'],
-      true,
-      false,
-    ),
-    createFileWithContent(
-      tokens,
-      mappedTokens,
-      tokenGroups,
-      '_radii.scss',
-      [TokenType.dimension],
-      ['Radius'],
-      true,
-      false,
-    ),
-    createFileWithContent(
-      tokens,
-      mappedTokens,
-      tokenGroups,
-      '_borders.scss',
-      [TokenType.dimension],
-      ['Border'],
-      false,
-      true,
-    ),
-    createFileWithContent(
-      tokens,
-      mappedTokens,
-      tokenGroups,
-      '_other.scss',
-      [TokenType.dimension, TokenType.string],
-      ['Grid', 'Container', 'Breakpoint'],
-      true,
-      true,
-    ),
-  ];
+      fileName,
+      tokenTypes,
+      groupNames,
+      withCssObject,
+      withParent,
+    );
+  });
 };
